+++ .idea/artifacts/BlogWeb_war.xml
@@ -1,0 +1,14 @@
+<component name="ArtifactManager">
+  <artifact type="war" name="BlogWeb:war">
+    <output-path>$PROJECT_DIR$/target</output-path>
+    <properties id="maven-jee-properties">
+      <options>
+        <module>BlogWeb</module>
+        <packaging>war</packaging>
+      </options>
+    </properties>
+    <root id="archive" name="BlogWeb-1.0.war">
+      <element id="artifact" artifact-name="BlogWeb:war exploded" />
+    </root>
+  </artifact>
+</component>
