--- src/java/com/blogapp/controller/Controller.java
@@ -1,30 +1,0 @@
-package com.blogapp.controller;
-
-import com.blogapp.controller.command.Command;
-import com.blogapp.controller.command.CommandProvider;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class Controller extends HttpServlet {
-
-    private static final String PARAMETER_COMMAND = "command";
-
-    private final CommandProvider provider = new CommandProvider();
-
-    public Controller() { super(); }
-
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String commandName = request.getParameter(PARAMETER_COMMAND);
-        Command command = provider.getCommand(commandName);
-
-        command.execute(request, response);
-    }
-
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        doGet(request, response);
-    }
-}
