--- src/java/com/blogapp/controller/command/Command.java
@@ -1,10 +1,0 @@
-package com.blogapp.controller.command;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public interface Command {
-    void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException;
-}
