--- src/java/com/blogapp/dao/ArticalDAO.java
@@ -1,9 +1,0 @@
-package com.blogapp.dao;
-
-import com.blogapp.bean.Article;
-
-public interface ArticalDAO {
-    void addArtical(Article article);
-    void deleteArtical(long idArtical);
-    void delete(Article article);
-}
