--- src/java/com/blogapp/dao/impl/SQLArticalDAO.java
@@ -1,22 +1,0 @@
-package com.blogapp.dao.impl;
-
-import com.blogapp.bean.Article;
-import com.blogapp.dao.ArticalDAO;
-
-public class SQLArticalDAO implements ArticalDAO{
-
-    @Override
-    public void addArtical(Article article) {
-
-    }
-
-    @Override
-    public void deleteArtical(long idArtical) {
-
-    }
-
-    @Override
-    public void delete(Article article) {
-
-    }
-}
