--- src/java/com/blogapp/service/BlogService.java
@@ -1,8 +1,0 @@
-package com.blogapp.service;
-
-import com.blogapp.bean.Article;
-
-public interface BlogService {
-    void addNewArtical(Article article);
-    void addEditedArtical(Article article);
-}
