--- src/java/com/blogapp/service/ReaderService.java
@@ -1,11 +1,0 @@
-package com.blogapp.service;
-
-import com.blogapp.bean.User;
-import com.blogapp.service.exception.ServiceException;
-
-public interface ReaderService {
-    User signIn(String login, byte[] password) throws ServiceException;
-    void signOut(String login);
-    void registration(User user);
-
-}
