--- src/java/com/blogapp/service/impl/BlogServiceImpl.java
@@ -1,17 +1,0 @@
-package com.blogapp.service.impl;
-
-import com.blogapp.bean.Article;
-import com.blogapp.service.BlogService;
-
-public class BlogServiceImpl implements BlogService{
-
-    @Override
-    public void addNewArtical(Article article) {
-
-    }
-
-    @Override
-    public void addEditedArtical(Article article) {
-
-    }
-}
