+++ src/main/java/com/blog/controller/CharsetFilter.java
@@ -1,0 +1,27 @@
+package com.blog.controller;
+
+import javax.servlet.*;
+import java.io.IOException;
+
+public class CharsetFilter implements Filter {
+
+    private static final String CONFIG_PARAM_ENCODING = "encoding";
+    private String encoding;
+
+    @Override
+    public void init(FilterConfig filterConfig) throws ServletException {
+        encoding = filterConfig.getInitParameter(CONFIG_PARAM_ENCODING);
+    }
+
+    @Override
+    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+        servletRequest.setCharacterEncoding(encoding);
+        servletResponse.setCharacterEncoding(encoding);
+        filterChain.doFilter(servletRequest, servletResponse);
+    }
+
+    @Override
+    public void destroy() {
+
+    }
+}
