+++ src/main/java/com/blog/controller/command/Command.java
@@ -1,0 +1,10 @@
+package com.blog.controller.command;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public interface Command {
+    void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException;
+}
