+++ src/main/java/com/blog/controller/command/impl/SignOut.java
@@ -1,0 +1,16 @@
+package com.blog.controller.command.impl;
+
+import com.blog.controller.command.Command;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public class SignOut implements Command {
+
+    @Override
+    public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+    }
+}
