+++ src/main/java/com/blog/dao/ArticalDAO.java
@@ -1,0 +1,9 @@
+package com.blog.dao;
+
+import com.blog.bean.Article;
+
+public interface ArticalDAO {
+    void addArtical(Article article);
+    void deleteArtical(long idArtical);
+    void delete(Article article);
+}
