+++ src/main/java/com/blog/dao/DAOProvider.java
@@ -1,0 +1,22 @@
+package com.blog.dao;
+
+import com.blog.dao.impl.SQLUserDAO;
+import com.blog.dao.impl.SQLArticalDAO;
+
+public class DAOProvider {
+    private static final DAOProvider instance = new DAOProvider();
+
+    private final UserDAO userDAO = new SQLUserDAO();
+    private final ArticalDAO articalDAO = new SQLArticalDAO();
+
+    private DAOProvider() {};
+
+    public static DAOProvider getInstance(){
+        return instance;
+    }
+
+    public UserDAO getUserDAO() { return userDAO; }
+
+    public ArticalDAO getArticalDAO() { return articalDAO; }
+
+}
