+++ src/main/java/com/blog/dao/exception/DAOException.java
@@ -1,0 +1,23 @@
+package com.blog.dao.exception;
+
+public class DAOException extends Exception {
+    public DAOException(String message, Exception e) {
+        super(message, e);
+    }
+
+    public DAOException() {
+        super();
+    }
+
+    public DAOException(String message) {
+        super(message);
+    }
+
+    public DAOException(Throwable cause) {
+        super(cause);
+    }
+
+    public DAOException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
