+++ src/main/java/com/blog/dao/impl/SQLArticalDAO.java
@@ -1,0 +1,22 @@
+package com.blog.dao.impl;
+
+import com.blog.bean.Article;
+import com.blog.dao.ArticalDAO;
+
+public class SQLArticalDAO implements ArticalDAO{
+
+    @Override
+    public void addArtical(Article article) {
+
+    }
+
+    @Override
+    public void deleteArtical(long idArtical) {
+
+    }
+
+    @Override
+    public void delete(Article article) {
+
+    }
+}
