+++ src/main/java/com/blog/service/BlogService.java
@@ -1,0 +1,8 @@
+package com.blog.service;
+
+import com.blog.bean.Article;
+
+public interface BlogService {
+    void addNewArtical(Article article);
+    void addEditedArtical(Article article);
+}
