+++ src/main/java/com/blog/service/ReaderService.java
@@ -1,0 +1,11 @@
+package com.blog.service;
+
+import com.blog.bean.User;
+import com.blog.service.exception.ServiceException;
+
+public interface ReaderService {
+    User signIn(String login, byte[] password) throws ServiceException;
+    boolean signOut(String login);
+    void registration(User user);
+
+}
