+++ src/main/java/com/blog/service/impl/BlogServiceImpl.java
@@ -1,0 +1,17 @@
+package com.blog.service.impl;
+
+import com.blog.bean.Article;
+import com.blog.service.BlogService;
+
+public class BlogServiceImpl implements BlogService{
+
+    @Override
+    public void addNewArtical(Article article) {
+
+    }
+
+    @Override
+    public void addEditedArtical(Article article) {
+
+    }
+}
