+++ src/main/webapp/index.jsp
@@ -1,0 +1,20 @@
+<%@ page contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Blog</title>
+</head>
+
+<header>
+    <jsp:include page="/WEB-INF/jsp/header.jsp"/>
+</header>
+
+<body>
+Blog
+</body>
+
+<footer>
+    <jsp:include page="/WEB-INF/jsp/footer.jsp"/>
+</footer>
+
+</html>
