--- webapp/index.jsp
@@ -1,21 +1,0 @@
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
-<%@ page contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
-<!DOCTYPE html>
-<html>
-<head>
-    <title>JSP - Hello World</title>
-</head>
-<body>
-<header>
-    <jsp:include page="WEB-INF/jsp/header.jsp"/>
-</header>
-
-<main>
-
-</main>
-
-<footer>
-    <jsp:include page="WEB-INF/jsp/footer.jsp"/>
-</footer>
-</body>
-</html>
